Moved some debug scripts and updated readme in learn2slide
