+++ scripts/debug/infinity_demo/infinity_demo_update.py
@@ -1,6 +1,6 @@
-## infinity_demo_hw.py
+## infinity_simulation.py
 #   Description:
-#       This function is meant to produce the infinity motion on our Kinova arm.
+#       This function is meant to produce the infinity motion in the meshcat simulator for our Kinova arm.
 #       We will use the comand sequence controller demonstrated in the peg_pickup_demo.py from kinova_drake
 ##
 
@@ -12,19 +12,9 @@ from pydrake.all import *
 import numpy as np
 import matplotlib.pyplot as plt
 
-# Start a single meshcat server instance to use for the remainder of this notebook.
-server_args = []
-from meshcat.servers.zmqserver import start_zmq_server_as_subprocess
-proc, zmq_url, web_url = start_zmq_server_as_subprocess(server_args=server_args)
-
-
-## import kinova drake which is not in the subdirectories here
-import sys
-sys.path.append('/root/kinova_drake/')
-
-from kinova_station import KinovaStationHardwareInterface, EndEffectorTarget, GripperTarget, KinovaStation
-from controllers import CommandSequenceController, CommandSequence, Command
-from observers.camera_viewer import CameraViewer
+from kinova_drake.kinova_station import KinovaStationHardwareInterface, EndEffectorTarget, GripperTarget, KinovaStation
+from kinova_drake.controllers import CommandSequenceController, CommandSequence, Command
+# from kinova_drake.observers.camera_viewer import CameraViewer
 
 ####################
 # Helper Functions #
@@ -109,11 +99,11 @@ def create_end_effector_target(ee_command_type,builder,station):
             station.GetInputPort("ee_target"))
 
 
-def create_infinity_hw_scenario(station):
+def create_infinity_scenario():
     """
     Description:
-        Creates the the systems and connections that work with the hardware Kinova station.
-        Understands that the Robotiq gripper 2f_85 is used.
+        Creates the 6 degree of freedom Kinova system in simulation. Anchors it to a "ground plane" and gives it the
+        RobotiQ 2f_85 gripper.
         This should also initialize the meshcat visualizer so that we can easily view the robot.
     Usage:
     """
@@ -122,12 +112,18 @@ def create_infinity_hw_scenario(station):
     gripper_type = '2f_85'
 
     # Start with the Kinova Station object
-    # station = KinovaStation(time_step=0.001,n_dof=6)
+    station = KinovaStation(time_step=0.001,n_dof=6)
 
+    # station.AddArm()
+    # station.Add2f85Gripper()
+    station.AddArmWith2f85Gripper()
+
+    station.AddGround()
+    station.AddCamera()
     # station.SetupSinglePegScenario(gripper_type=gripper_type, arm_damping=False)
-    # station.ConnectToMeshcatVisualizer(zmq_url="tcp://127.0.0.1:6000")
+    station.ConnectToMeshcatVisualizer()
 
-    #station.Finalize() # finalize station (a diagram in and of itself)
+    station.Finalize() # finalize station (a diagram in and of itself)
 
     # Start assembling the overall system diagram
     builder = DiagramBuilder()
@@ -150,11 +146,10 @@ def create_infinity_hw_scenario(station):
     diagram_context = diagram.CreateDefaultContext()
 
     # context = diagram.CreateDefaultContext()
-    # station.meshcat.load()
     diagram.Publish(diagram_context)
 
     # Return station
-    return builder, controller, diagram, diagram_context
+    return builder, controller, station, diagram, diagram_context
 
 def setup_infinity_command_sequence():
     """
@@ -162,14 +157,14 @@ def setup_infinity_command_sequence():
         Creates the command sequence that we need to achieve the infinity sequence.
     """
     # Constants
-    num_points_in_discretized_curve = 10
+    num_points_in_discretized_curve = 7
     infinity_center = np.array([0.6, 0.0, 0.5])
     infinity_center_pose = np.zeros((6,))
     infinity_center_pose[:3] = np.array([np.pi/2,0.0,0.5*np.pi])
     infinity_center_pose[3:] = infinity_center
 
     curve_radius = 0.2
-    curve_duration = 8.0
+    curve_duration = 6.0
 
     infinity_left_lobe_center = infinity_center - np.array([0.0,0.25,0.0])
     infinity_left_lobe_center_pose = np.zeros((6,))
@@ -249,8 +244,8 @@ def setup_controller_and_connect_to_station(cs,builder,station):
 
     # Create the controller and connect inputs and outputs appropriately
     #Kp = 10*np.eye(6)
-    Kp = np.diag([1,1,1,0.2,0.2,0.2])
-    Kd = 0.2*np.sqrt(Kp)
+    Kp = np.diag([10,10,10,2,2,2])
+    Kd = 2*np.sqrt(Kp)
 
     controller = builder.AddSystem(CommandSequenceController(
         cs,
@@ -276,20 +271,20 @@ show_toplevel_diagram = False
 # Which gripper to use (hande or 2f_85)
 gripper_type = "2f_85"
 
-n_dof = 6
+run = True
 
 ##########################
 # Setting Up the Station #
 ##########################
 
-with KinovaStationHardwareInterface(n_dof) as station:
+# Set up the kinova station
+builder, controller, station, diagram, diagram_context = create_infinity_scenario()
+#station.SetArmPositions(diagram,diagram_context,np.array([0.1,0.0,0.0,2.0,0.0,0.2]))
 
-    # Set up the kinova station
-    builder, controller, diagram, diagram_context = create_infinity_hw_scenario(station)
-    #station.SetArmPositions(diagram,diagram_context,np.array([0.1,0.0,0.0,2.0,0.0,0.2]))
+if run:
 
     # First thing: send to home position
-    station.go_home('Home')
+    station.go_home(diagram,diagram_context)
 
     # We use a simulator instance to run the example, but no actual simulation 
     # is being done: it's all on the hardware. 
@@ -305,7 +300,7 @@ with KinovaStationHardwareInterface(n_dof) as station:
 
     # Run simulation
     simulator.Initialize()
-    simulator.AdvanceTo(50.0)
+    simulator.AdvanceTo(40.0)
 
 while True:
     1
