+++ scripts/learning2slide/readme.md
@@ -1,0 +1,3 @@
+# learn2slide
+
+Project for sliding a car using the kinova arm.
